requirements.txt
pkg-resources==0.0.0
python-dotenv==0.10.3
selenium==3.141.0
urllib3==1.25.3
